Move e2e mocks to "e2e-tests" folder
