--- libs/transactions-journey/e2e-tests/fixture/mocks.ts
@@ -1,60 +1,0 @@
-import { Page } from '@playwright/test';
-import { readFileSync } from 'node:fs';
-
-const accessControlMocks = {
-  success: {
-    '**/accessgroups/users/permissions/summary': JSON.parse(
-      readFileSync(
-        './mock-server/mocks/golden-sample-app/accessgroups/users_permission_summary.json',
-        'utf-8'
-      )
-    ).responses.ok.data,
-    '**/accessgroups/service-agreements/context': JSON.parse(
-      readFileSync(
-        './mock-server/mocks/golden-sample-app/accessgroups/user-context_service-agreements.json',
-        'utf-8'
-      )
-    ).responses.ok.data,
-  },
-};
-const arrangementsMocks = {
-  success: {
-    '**/productsummary/context/arrangements**': JSON.parse(
-      readFileSync(
-        './mock-server/mocks/golden-sample-app/productsummary/context_arrangements.json',
-        'utf-8'
-      )
-    ).responses.ok.data,
-  },
-};
-const transactionsMocks = {
-  success: {
-    '**/v2/transactions': JSON.parse(
-      readFileSync(
-        './mock-server/mocks/golden-sample-app/transactions/transactions.json',
-        'utf-8'
-      )
-    ).responses.ok.data,
-  },
-};
-
-export const defaultListMocks = {
-  ...accessControlMocks.success,
-  ...arrangementsMocks.success,
-  ...transactionsMocks.success,
-};
-
-export const defaultDetailsMocks = {
-  ...accessControlMocks.success,
-  ...arrangementsMocks.success,
-  ...transactionsMocks.success,
-};
-
-export function setupPageMocks(
-  page: Page,
-  mocks: Record<string, string | object>
-) {
-  for (const [endpoint, data] of Object.entries(mocks ?? {})) {
-    page.route(endpoint, (route) => route.fulfill({ json: data }));
-  }
-}
