+++ libs/transactions-journey/e2e-tests/mocks/access-control-context.json
@@ -1,0 +1,15 @@
+[
+  {
+    "additions": {},
+    "id": "12345ed2abc2134",
+    "externalId": "sa_sdbxaz-ankit",
+    "name": "Ankit",
+    "description": "Master Service Agreement for Ankit",
+    "isMaster": true,
+    "validFromDate": null,
+    "validFromTime": null,
+    "validUntilDate": null,
+    "validUntilTime": null,
+    "customerCategory": null
+  }
+]
